[REFACTOR]: Dto로 변경 및 dto디렉토리로 이동
