+++ pickme/src/main/java/project/pickme/item/dto/ImageDto.java
@@ -1,11 +1,11 @@
-package project.pickme.item.domain;
+package project.pickme.item.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class ImageDTO {
+public class ImageDto {
 
 	private Long item_id;
 	private String name;
