+++ pickme/src/main/java/project/pickme/item/dto/LawDto.java
@@ -1,11 +1,11 @@
-package project.pickme.item.domain;
+package project.pickme.item.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class LawDTO {
+public class LawDto {
 	private Long id;
 	private String title;
 }
